+++ app/javascript/article-form/elements/__mocks__/algoliasearch.js
